action: simplify action_prompt_command()
