+++ src/meson.build
@@ -1,6 +1,5 @@
 labwc_sources = files(
   'action.c',
-  'action-prompt-command.c',
   'buffer.c',
   'debug.c',
   'desktop.c',
